+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.0-beta4 (29.02.2024)
+
+* Home screen menu arrangement
+* Create folder feature added
+* Minor UI fix
+
 ## v1.0.0-beta3 (16.02.2024)
 
 * UI improvement